+++ frontend/README.md
@@ -29,4 +29,4 @@ spellchecker https://streetsidesoftware.com/vscode-spell-checker/docs/configurat
 id중복검사, ip기준 hitcount,소켓쓰는 알림
 소셜로그인, 소셜공유, 검색(구글이용), 메세지,
 등등 기능들 더있갰지만 
-404,내정보,댓글,대댓글까지 구현하고 배포할 예정
+내정보,댓글,대댓글까지 구현하고 배포할 예정
